Storybook: do not use separate theme provider, just use matchMedia
The previous dropdown selection wasn't working perfectly since we often use css media queries to make dark / light changes. It's best to use devTools or browser/OS settings to switch between
